requirements.txt
streamlit
google-generativeai
pypdf2
python-docx
pandas
python-dotenv
openpyxl
